Fix #1 - Esimated Operator Cost Incorrect
Estimated operator cost needs to be calculated as the EstimatedTotalSubtreeCost for that RelOp minus the EstimatedTotalSubtreeCost for all child nodes, with a minimum value of 0.

This can be fiddly to calculate as the EstimatedTotalSubtreeCost can be a scientific number. This solution uses a fragment variable and node-set - it also needs to check to see which of the `exslt:node-set` and `msxml:node-set` functions are available to make sure it works cross-browser.

Also fixed newlines for the "Estimated Operator Cost" tooltip row, as it makes unit testing easier.