Fix #12 - Change canvas positioning
Change the canvas to be 100% width / height of .qp-root, and fix .qp-root to size to its contents.

Canvas is added to .qp-root instead of the container as it should give more flexibility on what CSS can be used with the container while still having the lines appear correctly.

Also moved some positioning CSS from JavaScript into stylesheet.