Fix cost percentages incorrect for cursors.
Looks like the XSLT was showing the cost percentage for the first `s:RelOp`. I'm yet to see a plan with costs other than 0% for cursors, so hard-coding this for now.

I think I could do the same for `s:Stmtimple`, but I'm not 100% confident at the moment so I'll re-visit this.