Don't hide tooltip while mouse over tooltip.
Tooltip has nested children, so need to use findAncestor to identify whether the element raising the event is a descendant of either the node or its tooltip.

Also need to handle the mouseout event of the tooltip itself as otherwise the tooltip may continue to be displayed in some cases.