#53 - Show missing indexes
Missing indexes are shown underneath the query text in green. The CSS for these header rows needed adjusting to make sure that all rows (the query text and the indexes) don't just appear on top of each other, as each row needs to be absolutely positioned to ensure that the text is truncated properly, that all batches in the supplied XML are shown with the same width.