Start refactor ToolTipDetails to use call-template
The template for the "Warnings" section needs access to the s:MissingIndexes element, which is unfortunately on the parent node if we match by s:Warnings. I can change the template to match on `*[s:Warnings]`, but then because we also match on */* when applying ToolTipDetails this means that we actually match on the previous node as well. I tried refacforing so that we don't need to match on */*, but that caused all sorts of issues.

The short version is that using apply-templates when calling `ToolTipDetails` isn't a great idea, so I'm going to slowly refactor to using call-template instead. Its too big a change to do all at once, so  I'm just going to use apply-template as well, as call-template and slowly refactor templates into the latter.