Fix #61 - Add unmatched indexes to tooltip
The mode="QpTr" template matches on either s:RelOp or s:Stmt... nodes, however in the case of s:RelOp the s:Warnings element is a direct child but for s:Stmt... the s:Warnings element is actually a child of s:QueryPlan.

For now I've worked around this by adding a hack to change context to the s:QueryPlan element, but a better solution would be to see if mode="QpTr" template should instead match on the s:QueryPlan element instead.