Improve test for queryPlan property
Was hoping to avoid parsing the XML twice but in IE to run XSLT we need to parse the XML in a way which works totally differently from other browsers. To solve this I parse the XML twice, once just for the XSLT (which works differently in IE), and then again using DOMParser so the XML is accessible in a consistent way across both browsers.