Setup CSS so that SVG lines recieve pointer events
Used to use z-index to put the svg behind everything else so it doesn't prevent tooltips from being shown for nodes, but now we want this layer to see some events (but still not obscure nodes).

To solve this we give the qp-root pointer-events none so the canvas itself can be clicked through, but then each line is given pointer-events auto so their tooltips still work.

This solution still needs checking in Firefox & IE.